// packages/beautify/src/types/prettydiff.ts
interface EnforcedRules {

  /**
   * We only want the `beautify` feature from PrettyDiff. So
   * we will assert and enforce this `mode` for all rules.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * The operation to be performed.
   *
   * **PrettyDiff Options**
   *
   * > `beautify`
   *  Beautifies code and returns a string (**selected**)

   * > ~~`diff`~~
   *  Returns either command line list of differences or an HTML report
   *
   * > ~~`minify`~~
   *  Minifies code and returns a string
   *
   * > ~~`parse`~~
   * Using option `parseFormat` returns an object with shallow arrays,
   * a multidimensional array, or an HTML report.
   *
   */
  readonly mode: 'beautify',

  /**
   * We never want this rule active, it is not required in Liquify
   * or an LSP environment.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * An educational tool to generate HTML output of JavaScript
   * code to identify scope regions and declared references by color.
   * This option is ignored unless the code language is JavaScript or TypeScript.
   *
   * **PrettyDiff Options**
   *
   * > ~~`html`~~
   *  Generates HTML output with escaped angle braces and ampersands for
   *  embedding as code, which is handy in code producing tools
   *
   * > `none`
   *  Prevents use of this option (**selected**)
   *
   * > ~~`report`~~
   *  Generates HTML output that renders in web browsers
   *
   */
  readonly jsscope: 'none'

  /**
   * We will not activate this rule. No terminal logging is supported in
   * LSP or Liquify, logs are passed directly to the editor.
   *
   * ---
   *
   *  **PrettyDiff Description**
   *
   * A node only option to determine if terminal summary data should be
   * logged to the console.
   *
   * **PrettyDiff Options**
   *
   * > ~~`double`~~
   *  Default minimal summary
   *
   * > `log`
   *  Verbose logging (**selected**)
   *
   * > ~~`quiet`~~
   *  No extraneous logging
   *
   */
  readonly end_quietly: 'quiet';

  /**
   * We enable this rule, unsure of its actual use-case but might be related
   * to error logging when parse issues occur.
   *
   * ---
   *
   *  **PrettyDiff Description**
   *
   * A Node.js only option if parse errors should be written to the console.
   */
  readonly node_error: true;

  /**
   * This will insert characters instead of whitespaces for indentation. This
   * is **NEVER** desirable within this environment.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * The string characters to comprise a single indentation. Any string
   * combination is accepted.
   */
  readonly indent_char: ' '
}

interface SharedRules {
  /**
   * PrettyDiff defaults this to `4` but we overwrite to `2` or assign
   * the vscode workspace editor options.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * The number of `indent_char` values to comprise a single indentation.
   *
   */
  indent_size: number,

  /**
   * When no `indent_level` is defined in a `.liquirc` file, then indentation
   * levels are inhertied from the Text Editors workspace settings. If
   * an `.editorconfig` file is found present in root, those rules will be
   * applied in **precedence** over Text Editor.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * How much indentation padding should be applied to beautification?
   * This option is internally used for code that requires switching
   * between libraries.
   *
   * @default 0
   */
  indent_level: number,

  /**
   * When no `wrap` is defined in a `.liquirc` file, then wrap levels are
   * inhertied from the Text Editors workspace settings.
   *
   * **Notes**
   *
   * When using with `force_attribute` Liquid singular (output) type
   * tags will format inline with tag pairs. This is an issue that needs
   * to be fixed.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * Character width limit before applying word wrap. A `0` value
   * disables this option. A negative value concatenates script strings.
   *
   * @default 0
   */
  wrap: number

  /**
   * PrettyDiff defaults this to `0` but we overwrite to `2` or assign
   * the vscode workspace editor options.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * The maximum number of consecutive empty lines to retain
   *
   * @default 2
   */
  preserve: number,

  /**
   * Whether or not to insert a final line. When this rule is undefined in
   * a `.liquidrc` file the Text Editors settings will be used, in vscode
   * that is `*.endWithNewline` where `*` is a language name.  If an
   * `.editorconfig` file is found present in root, those rules will be
   * applied in **precedence** over Text Editor.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * Insert an empty line at the end of output.
   *
   * @default false
   */
  new_line: boolean
}

export interface HTMLEnforce {

  readonly language_name: 'HTML/Liquid';
  readonly language: 'html';
  readonly language_default: 'html';
  readonly lexer: 'markup';

  /**
   * We want to avoid this at all costs, it should ALWAYS be `false`.
   * It's a legacy PrettyDiff rule that will wreak havoc within HTML.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * Allows immediately adjacement start and end markup tags
   * of the same name to be combined into a single self-closing tag.
   */
  readonly tag_merge: false

  /**
   * We want to avoid this at all costs, it should ALWAYS be `false`.
   * It's attempts to sort tag placements. Its wreaks havoc within HTML.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * Sort child items of each respective markup parent element.
   */
  readonly tag_sort: false
}

export interface HTMLRules extends SharedRules {

  /**
   * HTML Attribute sorting. When enabled it will sort attributes
   * alphabetically.
   *
   * ---
   *
   * **PrettyDiff Description**
   *
   * Alphanumerically sort markup attributes. Attribute sorting is
   * ignored on tags that contain attributes template attributes.
   *
   * @default false
   */
  attribute_sort: boolean

  /**
   * Define a sort list from which attributes should be arranged
   *
   * ---
   *
   *  **PrettyDiff Description**
   *
   * A comma separated list of attribute names. Attributes will be sorted according to
   * this list and then alphanumerically. This option requires 'attribute_sort' have
   * a value of true.
   *
   * @default ''
   */
  attribute_sort_list: string

  /**
   * **PrettyDiff Description**
   *
   * If a blank new line should be forced above comments.
   *
   * @default false
   */
  comment_line: boolean

  /**
   * **PrettyDiff Description**
   *
   * This will determine whether comments should always start at position
   * `0` of each line or if comments should be indented according to the code.
   *
   * @default false
   */
  comments: boolean

  /**
   * **PrettyDiff Description**
   *
   * Insert an empty line at the end of output
   *
   * @default false
   */
  new_line: boolean,

  /**
   * **PrettyDiff Description**
   *
   * Markup self-closing tags end will end with `' />'` instead of `'/>'`
   *
   * @default false
   */
  space_close: boolean,

  /**
   * **PrettyDiff Description**
   *
   * Prevent comment reformatting due to option wrap.
   *
   * @default true
   */
  preserve_comment: boolean,

  /**
   * **PrettyDiff Description**
   *
   * If text in the provided markup code should be preserved exactly as provided.
   * This option eliminates beautification and wrapping of text content.
   *
   * @default false
   */
  preserve_text: boolean,

  /**
   * **PrettyDiff Description**
   *
   * Automatically correct some sloppiness in code
   *
   * @default false
   */
  correct: boolean

  /**
   * **PrettyDiff Description**
   *
   * If all markup attributes should be indented each onto their own line._
   *
   * @default false
   */
  force_attribute: boolean

  /**
   * **PrettyDiff Description**
   *
   * Will force indentation upon all content and tags without regard for the
   * of new text nodes.
   *
   * @default false
   */
  force_indent: boolean

  /**
   * **PrettyDiff Description**
   *
   * If the quotes of markup attributes should be converted to single quotes
   * or double quotes.
   *
   * **PrettyDiff Options**
   *
   * - `double` Converts single quotes to double quotes
   * - `none` Ignores this option
   * - `single` Converts double quotes to single quotes
   *
   * ---
   *
   * @default 'none'
   */
  quote_convert: 'double' | 'single' | 'none'

  /**
   * **PrettyDiff Description**
   *
   * If markup tags should have their insides preserved.
   * This option is only available to markup and does not support
   * child tokens that require a different lexer.
   *
   * @default false
   */
  unformatted: boolean

}

export interface IHTMLRules extends EnforcedRules, HTMLEnforce, HTMLRules {}